requirements.txt
streamlit
langchain
langchain-community
langchain-ollama
pydantic
python-dotenv
python-docx
pypdf
openpyxl